Cleanup UX slightly
